guessing game
We start by importing the necessary classes: Random for generating random numbers and Scanner for reading user input.

We create a Random object called random to generate a random number.

We use the nextInt() method of the Random class to generate a random number between 1 and 100. We add 1 to the result to ensure that the number is inclusive of both 1 and 100. We store this random number in the variable randomNumber.

We declare three variables: guess to store the user's guess, attempts to keep track of the number of attempts, and scanner to read user input.

We display a welcome message and inform the player that a random number has been generated.

We start a do-while loop that will continue until the player guesses the correct number.

Inside the loop, we prompt the player to enter their guess using the print() method of System.out. We then read the user's input using the nextInt() method of the Scanner class and store it in the guess variable.

We increment the attempts variable to keep track of the number of attempts.

We use an if-else statement to check if the guess is too low, too high, or correct. If the guess is too low, we display a message indicating that the guess was too low. If the guess is too high, we display a message indicating that the guess was too high. If the guess is correct, we display a congratulatory message along with the number of attempts it took to guess the correct number.

The loop continues until the player guesses the correct number. Once the correct number is guessed, the loop terminates.

Finally, we close the Scanner object to release any system resources associated with it.

That's it! You have successfully created a simple guessing game in Java. Have fun playing and experimenting with the code!